requirements.txt
Faker==38.2.0
mysql-connector-python==9.5.0
tzdata==2025.2
